from int to float coordinates

+++ sql/user/user.py
@@ -14,8 +14,8 @@ class UserTable(db.Model):
     # For target
     age = db.Column(db.String(255))
     salary = db.Column(db.String(255))
-    city_longitude = db.Column(db.Integer)
-    city_latitude = db.Column(db.Integer)
+    city_longitude = db.Column(db.Float)
+    city_latitude = db.Column(db.Float)
     city_name = db.Column(db.String(255))
     #
     appeals = db.Column(db.Integer, default = 0)
